backend/pyproject.toml
[project]
name = "autofighter"
version = "0+46b3b6ae6dff630fbeea9172522e407b8c38867d2830c4c66e3da7761f88fa0e"
description = "Quart backend for Midori AI AutoFighter"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "pillow>=11.3.0",
    "rich>=13.0.0",
    "sqlcipher3-binary>=0.5.4",
    "quart>=0.19.6",
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "cryptography>=41.0.0",
]
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["autofighter", "plugins"]

[project.optional-dependencies]
llm-cpu = [
    "torch",
    "torchaudio",
    "torchvision",
    "chatterbox-tts",
    "openai",
    "langgraph",
    "langchain-core",
    "langchain-openai",
    "langchain-ollama",
    "langchain-localai",
    "langchain-community",
    "langchain-chroma",
    "langchain-huggingface",
    "langchain-text-splitters",
    "pillow",
    "transformers",
    "sentence-transformers",
    "accelerate",
]

llm-cuda = [
    "torch",
    "torchaudio",
    "torchvision",
    "chatterbox-tts",
    "openai",
    "langgraph",
    "langchain-core",
    "langchain-openai",
    "langchain-ollama",
    "langchain-localai",
    "langchain-community",
    "langchain-chroma",
    "langchain-huggingface",
    "langchain-text-splitters",
    "pillow",
    "transformers",
    "sentence-transformers",
    "accelerate",
]

llm-amd = [
    "torch",
    "torchaudio",
    "torchvision",
    "chatterbox-tts",
    "openai",
    "langgraph",
    "langchain-core",
    "langchain-openai",
    "langchain-ollama",
    "langchain-localai",
    "langchain-community",
    "langchain-chroma",
    "langchain-huggingface",
    "langchain-text-splitters",
    "pillow",
    "transformers",
    "sentence-transformers",
    "accelerate",
]
